combine output helpers
