+++ src/vs/workbench/contrib/notebook/browser/viewModel/cellOutputTextHelper.ts
@@ -3,9 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { IClipboardService } from '../../../../../platform/clipboard/common/clipboardService.js';
+import { ILogService } from '../../../../../platform/log/common/log.js';
 import { NotebookTextModel } from '../../common/model/notebookTextModel.js';
 import { IOutputItemDto } from '../../common/notebookCommon.js';
-import { ICellViewModel } from '../notebookBrowser.js';
+import { ICellOutputViewModel, ICellViewModel } from '../notebookBrowser.js';
 
 interface Error {
 	name: string;
@@ -67,3 +69,38 @@ export function getOutputText(mimeType: string, buffer: IOutputItemDto) {
 	return text;
 }
 
+export async function copyCellOutput(mimeType: string | undefined, outputViewModel: ICellOutputViewModel, clipboardService: IClipboardService, logService: ILogService) {
+	const cellOutput = outputViewModel.model;
+	const output = mimeType && TEXT_BASED_MIMETYPES.includes(mimeType) ?
+		cellOutput.outputs.find(output => output.mime === mimeType) :
+		cellOutput.outputs.find(output => TEXT_BASED_MIMETYPES.includes(output.mime));
+
+	mimeType = output?.mime;
+
+	if (!mimeType || !output) {
+		return;
+	}
+
+	const text = getOutputText(mimeType, output);
+
+	try {
+		await clipboardService.writeText(text);
+
+	} catch (e) {
+		logService.error(`Failed to copy content: ${e}`);
+	}
+}
+
+export const TEXT_BASED_MIMETYPES = [
+	'text/latex',
+	'text/html',
+	'application/vnd.code.notebook.error',
+	'application/vnd.code.notebook.stdout',
+	'application/x.notebook.stdout',
+	'application/x.notebook.stream',
+	'application/vnd.code.notebook.stderr',
+	'application/x.notebook.stderr',
+	'text/plain',
+	'text/markdown',
+	'application/json'
+];
